+++ lib/App.ts
@@ -37,9 +37,10 @@ import {type TriFrostBodyParserOptions} from './utils/BodyParser/types';
 import {type LazyInitFn} from './utils/Lazy';
 import {RouteTree} from './routing/Tree';
 import {extractDomainFromHost} from './utils/Http';
-import {type createCss, type createScript} from './modules';
+import {type createScript} from './modules';
 import {mount as mountCss} from './modules/JSX/style/mount';
 import {mount as mountScript} from './modules/JSX/script/mount';
+import {type CssInstance} from './modules/JSX/style/use';
 
 const RGX_RID = /^[a-z0-9-]{8,100}$/i;
 
@@ -105,7 +106,7 @@ type AppOptions <Env extends Record<string, any>> = {
     trustProxy?:boolean;
     client?: {
         script?: ReturnType<typeof createScript>['script'];
-        css?: ReturnType<typeof createCss>;
+        css?: CssInstance<any, any, any, any>;
     };
 }
 
@@ -144,7 +145,7 @@ class App <
     #cache:TriFrostCache<Env>;
 
     /* Client-css instance */
-    #css:ReturnType<typeof createCss>|null = null;
+    #css:CssInstance<any, any, any, any>|null = null;
 
     /* Client-script instance */
     #script:ReturnType<typeof createScript>['script']|null = null;